+++ src/main/java/dotblueshoes/fogger/Fogger.java
@@ -5,43 +5,50 @@ import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.Mod.Instance;
 import net.minecraftforge.fml.common.Mod;
-
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraft.init.Blocks;
 
 import org.apache.logging.log4j.Logger;
 
-import dotblueshoes.fogger.config.ConfigHandler;
+import dotblueshoes.fogger.dependency.*;
+import dotblueshoes.fogger.config.*;
 import dotblueshoes.fogger.event.*;
 
 @Mod( modid = Fogger.MODID,  version = Fogger.VERSION, useMetadata = true)
 public class Fogger {
     public static final String MODID = "fogger";
-    public static final String VERSION = "2";
+    public static final String VERSION = "3";
 
     @Instance(MODID)
     public static Fogger instance;
-    //private static Logger logger;
+    private static Logger logger;
 
     @EventHandler
     public void preInitialize(FMLPreInitializationEvent event) {
-        //logger = event.getModLog();
-        ConfigHandler.initConfig(event);
+        logger = event.getModLog();
+        ConfigHandler.initConfig(event.getSuggestedConfigurationFile());
     }
 
     @EventHandler
     public void initialize(FMLInitializationEvent event) {
-        //MinecraftForge.EVENT_BUS.register(new ConfigHandler());
+        // MinecraftForge.EVENT_BUS.register(new ConfigHandler()); - GUI thingy
+
+        // Registring the presence of DynamicSurroundings Mod.
+        DynamicSurroundingsDependency.checkPresence();
+
+        // Registring Fog Event.
         MinecraftForge.EVENT_BUS.register(new FogHelper());
-        if (ConfigHandler.isFogGlobal == false) 
+        if (ConfigHandler.isFogGlobal == false) {
             MinecraftForge.EVENT_BUS.register(new FogEvent(ConfigHandler.getFogDefinitions(), ConfigHandler.getFogMapDefinitions()));
-        else 
+        } else 
             MinecraftForge.EVENT_BUS.register(new GlobalFogEvent());
     }
 
-    // public static void logInfo(String msg) {
-    //     final String PREFIX = "$$$###$$$###";    
-    //     logger.info(PREFIX + msg);
-    // }
+    // this.class.getName();
+
+    public static void logInfo(String msg) {
+        final String PREFIX = "INFO: ";
+        logger.info(PREFIX + msg);
+    }
 
 }