+++ src/main/java/dotblueshoes/fogger/dependency/SereneSeasonsDependency.java
@@ -1,0 +1,7 @@
+package dotblueshoes.fogger.dependency;
+
+import java.lang.reflect.*;
+
+public class SereneSeasonsDependency {
+
+}